// backend/src/common/enums/work-order-status.enum.ts
export enum WorkOrderStatus {
  PENDING = 'pending',
  IN_PROGRESS = 'in_progress',
  WAITING_MATERIAL = 'waiting_material',
  WAITING_APPROVAL = 'waiting_approval',
  TRANSFERRING = 'transferring',
  REFUND_NEGOTIATING = 'refund_negotiating',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
  REFUNDED = 'refunded',
}